refactor drive module
